+++ packages/agility_redux_bloc/CHANGELOG.md
@@ -7,3 +7,17 @@
 ## [0.0.6] - fix rebuild
 
 ## [0.0.7] - BlocPage BlocPopupBox add animation
+
+## [0.0.8] - add NavigatorObserver
+
+## [0.0.9] - add Page arguments: disableAnimate
+
+## [0.0.10] - add pushOrReplacement function
+
+## [0.0.11] - fix pop state bug
+
+## [0.0.12] - add bloc widget child
+
+## [0.0.13] - update agility_redux_widget
+
+## [0.0.14] - fix debug mode